adds passing Clothing tests

+++ src/main/java/models/items/Clothing.java
@@ -4,7 +4,7 @@ public class Clothing extends Item {
 
     private Size size;
 
-    public Clothing(String name, int price, String description, int quantity, Size size) {
+    public Clothing(String name, double price, String description, int quantity, Size size) {
         super(name, price, description, quantity);
         this.size = size;
     }
